Split API routes into session vs non-session
